fix: remove cronjob
