+++ apps/studio/src/server/bullmq/queues/index.ts
@@ -13,7 +13,7 @@ export const REMOVE_ON_FAIL_BUFFER = 60 * 60 * 24 * 4 // 4 days in seconds
 export const WORKER_CONCURRENCY = 5
 export const WORKER_RETRY_LIMIT = 3
 export const WORKER_SHUTDOWN_TIMEOUT = 30_000 // 30 seconds
-const BACKOFF: BackoffOptions = {
+export const BACKOFF: BackoffOptions = {
   type: "exponential",
   delay: 10_000, // 10s, then 20s, then 40s etc
 }
@@ -41,22 +41,3 @@ export const REDLOCK_SETTINGS: Partial<Settings> = {
   retryDelay: 0,
   retryJitter: 0,
 }
-
-/** Job settings to be used by the caller */
-export const getJobOptionsFromScheduledAt = (
-  resourceId: string,
-  scheduledAt: Date,
-): JobsOptions => {
-  const delayInMs = scheduledAt.getTime() - Date.now()
-  return {
-    attempts: WORKER_RETRY_LIMIT,
-    backoff: BACKOFF,
-    removeOnComplete: { age: REMOVE_ON_COMPLETE_BUFFER },
-    removeOnFail: { age: REMOVE_ON_FAIL_BUFFER },
-    delay: delayInMs,
-    jobId: getJobIdFromResourceId(resourceId),
-  }
-}
-
-export const getJobIdFromResourceId = (resourceId: string) =>
-  `resource-${resourceId}`
