+++ apps/studio/src/server/cron/index.ts
@@ -1,8 +1,5 @@
 import { createBaseLogger } from "../../lib/logger"
-import {
-  cleanupScheduledPublishesJob,
-  deactivateInactiveUsersJob,
-} from "./jobs"
+import { deactivateInactiveUsersJob } from "./jobs"
 
 const logger = createBaseLogger({ path: "cron:index" })
 
@@ -14,7 +11,6 @@ export const initializeCronJobs = () => {
 
   // Initialize and track all cron jobs
   cronJobs.push(deactivateInactiveUsersJob())
-  cronJobs.push(cleanupScheduledPublishesJob())
 
   logger.info("Cron jobs initialized successfully")
 }
