+++ apps/studio/src/server/cron/jobs/index.ts
@@ -1,2 +1,1 @@
 export { deactivateInactiveUsersJob } from "./deactivateInactiveUsersJob"
-export { cleanupScheduledPublishesJob } from "./cleanupScheduledPublishesJob"
